+++ CHANGELOG.md
@@ -15,7 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - added tenant export, import and clone functions to myems-api, myems-admin
 - added offline meter export, import and clone functions to myems-api, myems-admin
 - added virtual meter export, import and clone functions to myems-api, myems-admin
-- added space export, import and clone functions to myems-api
+- added space export, import and clone functions to myems-api, myems-admin
 - added command export, import and clone functions to myems-api
 - added energy flow diagram export, import and clone functions to myems-api
 - added tariff export, import and clone functions to myems-api